backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
feedparser>=6.0
pyyaml>=6.0
pydantic>=2.10
